% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_annotate.R
\name{annotate_server}
\alias{annotate_server}
\alias{annotate_ui}
\title{Annotate clades with a name or color}
\usage{
annotate_server(id, react_list, plot_inputs)

annotate_ui(id)
}
\arguments{
\item{id}{namespace id}

\item{react_list}{\code{reaciveVal}. A list of variables updated within the app and passed between modules.}

\item{plot_inputs}{\code{reactive}. A list of inputs from an interactive plot. For now only plot brush (see \code{?shiny::plotOutput}) is handled.}
}
\description{
Annotate clades with a name or color
}
