% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_draw.R
\name{draw_server}
\alias{draw_server}
\title{Shiny module to draw a phylogeny}
\usage{
draw_server(id, phylogeny)
}
\arguments{
\item{id}{character, namespace id}

\item{phylogeny}{object of class `phylo` to draw}
}
\value{
UI and server logic for plotting and interactive modification of a phylogeny
}
\description{
Shiny module to draw a phylogeny
}
\details{
Trees with > 200 tips will in most cases be illegible. Tip names become illegible in smaller trees.
}
\examples{
if (interactive()) {

  ui <- fluidPage(
    titlePanel("Shiny module to draw phylogenetic trees"),
    mainPanel(
    draw_ui(id = "id")
    )
  )
  
  server <- function(input, output, session) {
    X <- ape::rtree(150)
    draw_server(id = "id", phylogeny = X)
  }
  
  shinyApp(ui = ui, server = server)
}

}
