% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_objects.R
\docType{data}
\name{palettes}
\alias{palettes}
\title{A list of color palettes}
\format{
Simple named list with color palettes (n=9) including all
\code{viridis} palettes and diverging, colorblind safe \code{RColorBrewer} palettes:
\describe{
\item{viridis}{}
\item{magma}{}
\item{inferno}{}
\item{plasma}{}
\item{cividis}{}
\item{BrBG}{}
\item{PiYG}{}
\item{PRGn}{}
\item{PuOr}{}
\item{RdBu}{}
\item{RdYlBu}{}
}
}
\usage{
palettes
}
\description{
A list of color palettes
}
\keyword{datasets}
