% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{show_modules}
\alias{show_modules}
\alias{srv_list}
\alias{uis_list}
\alias{layout_ui}
\alias{layout_server}
\alias{branches_ui}
\alias{branches_server}
\alias{tips_ui}
\alias{tips_server}
\alias{plotarea_ui}
\alias{plotarea_server}
\alias{misc_ui}
\alias{misc_server}
\title{Modules to customize the appearence of a phylogeny}
\usage{
show_modules()

srv_list()

uis_list()

layout_ui(id)

layout_server(id)

branches_ui(id)

branches_server(id)

tips_ui(id)

tips_server(id)

plotarea_ui(id)

plotarea_server(id)

misc_ui(id)

misc_server(id)
}
\arguments{
\item{id}{namespace id}
}
\description{
\strong{\code{show_modules()}}

Show a data frame of the available Shiny modules.

\strong{\code{srv_list()/uis_list()}}

Simple functions that return the available server and ui
components as named lists. These are used internally in \code{make_phynotate} to
generate the application UI and server.

\strong{\code{layout_ui()/layout_server()}}

Change the layout (rectangular, slanted, circular) and directon (left, right, ...)
of a phylogeny displayed through the draw module. See \code{?make_phynotate} for
example usage.

\strong{\code{branches_ui()/branches_server()}}

Change the appearence of the branches (line color, weight)
of a phylogeny displayed through the draw module. See \code{?make_phynotate} for
example usage.

\strong{\code{tips_ui()/tips_server()}}

Enable and change the appearence of the tip labels (font color, font size, offset)
of a phylogeny displayed through the draw module. See \code{?make_phynotate} for
example usage.

\strong{\code{plotarea_ui()/plotarea_server()}}

Change the plotting areaof a phylogeny displayed through the draw module.
For now, only plot height and width can be modified. See \code{?make_phynotate} for
example usage.

\strong{\code{misc_ui()/misc_server()}}

Less common customizations of a phylogeny displayed through the draw module.
For example, the to modify the amount of space (degrees) between the first and last tip
of a phylogeny plotted in circular layout. See \code{?make_phynotate} for
example usage.
}
