% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_phynotate.R
\name{make_phynotate}
\alias{make_phynotate}
\title{Make a Shiny application to display, customize appearence, and annotate a phylogeny}
\usage{
make_phynotate(phylogeny, modules = "LPTB")
}
\arguments{
\item{phylogeny}{an object of class \code{phylo} to display. Very large phylogenies
typically will be dificult to display legibly.}

\item{modules}{character string specifying the customizing widgets (shiny modules) to
include. The default is \code{"LPBT"} which stands for Layout, Plot area,
Branches, and Tips. Changing the content and order can be achieved by
changing the string, e.g., \code{"LP"} will only allow customization
of the plot layout and size, whereas \code{"BTL"} will show widgets for
branches, tips, and layout in that order.

See \code{?modules} for details about the modules.}
}
\description{
\code{make_phynotate} generates a standalone Shiny app with one phylogeny and
widgets for customizing its appearence. This is a wrapper around
\code{phynotate::draw_ui/server} to simplify usage when a the draw module is
called as a standalone app. See \code{?draw_server} for an example of how to include
the module in an existing app.
}
\examples{
if (interactive()) {
   library(ape)
   data(bird.orders)
   make_phynotate(phylogeny=bird.orders, modules = "LPBT") 
}

}
